requirements.txt
fastapi
uvicorn
supabase-py
python-jose
pydantic
python-dotenv
sqlalchemy
passlib
python-dateutil
typing-extensions
starlette
requests